Return changed to HttpResponseMessage. This allows a query of the StatusCodes and can pass useful messages to the user via try/catch

+++ Controllers/AlarmController.cs
@@ -1,3 +1,4 @@
+using System.Net;
 using BotAPI.Models;
 using BotAPI.Utility;
 using Microsoft.AspNetCore.Mvc;
@@ -15,27 +16,34 @@ public class AlarmController : ControllerBase
     {
         const int maxAlarm = 5;
         
-        Task<string> taskString = SvcConnector.SvcGetAsync("ActiveAlarm");
-        string responseString = taskString.Result;
-        var alarms = JsonConvert.DeserializeObject<List<AlarmDto>>(responseString);
-        var alarmCount = alarms.Count;
-        var response = new ResponseAlarm();
-        
-        //Sortiert die Alarme nach Datum(neuste zuerst)
-        alarms.Sort();
-        alarms.Reverse();
-        
-        //entfert alle Alarme die maxAlarm überschreiten
-        if (alarmCount > maxAlarm)
-        {
-            alarms.RemoveRange(maxAlarm, alarmCount - maxAlarm);
-        }
-        
-        //setzt die Alarm Liste zusammen
-        response.Alarms = alarms;
-        response.size = alarmCount;
-        response.name = "Active Alarms";
+        Task<HttpResponseMessage> response = SvcConnector.SvcGetAsync("ActiveAlarm");
+        Task<string> responseString = response.Result.Content.ReadAsStringAsync();
+        var convertString = responseString.Result;
 
-        return response;
+        if (response.Result.StatusCode == HttpStatusCode.OK)
+        {
+            var alarms = JsonConvert.DeserializeObject<List<AlarmDto>>(convertString);
+            var alarmCount = alarms.Count;
+            var responseAlarm = new ResponseAlarm();
+
+            //Sortiert die Alarme nach Datum(neuste zuerst)
+            alarms.Sort();
+            alarms.Reverse();
+
+            //entfert alle Alarme die maxAlarm überschreiten
+            if (alarmCount > maxAlarm)
+            {
+                alarms.RemoveRange(maxAlarm, alarmCount - maxAlarm);
+            }
+
+            //setzt die Alarm Liste zusammen
+            responseAlarm.Alarms = alarms;
+            responseAlarm.size = alarmCount;
+            responseAlarm.name = "Active Alarms";
+
+            return responseAlarm;
+        }
+
+        return BadRequest(responseString);
     }
 }
